refactor: Enhance DoOrder component with ComboMeal support and cart management improvements - Introduced CartItemType enum to differentiate between menu items and combo meals. - Updated addToCart function to handle both MenuItem and ComboMeal types. - Enhanced order submission logic to include item type in the order data. - Improved UI to allow selection between menu items and combo meals.
